Former-commit-id: ce00fe1443dcd95c280eda746381421d80be779d

+++ ConsoleApplication/HortaTracer/src/org/janelia/horta/ImmutableNeuriteAnchor.java
@@ -42,7 +42,7 @@ public interface ImmutableNeuriteAnchor
     
     public ConstVector3 getLocationUm();
 
-    public int getIntensity();
+    public double getIntensity();
 
     public float getRadiusUm();
 }
